src/components/instance-page/instance-page.spec.ts
import { flush, render } from '@stencil/core/testing';
import { InstancePage } from './instance-page';

describe('instance-page', () => {
  it('should build', () => {
    expect(new InstancePage()).toBeTruthy();
  });

  describe('rendering', () => {
    let element;
    beforeEach(async () => {
      element = await render({
        components: [InstancePage],
        html: '<instance-page></instance-page>'
      });
    });

    it('should not render any content if there is not a match', async () => {
      await flush(element);
      expect(element.textContent).toEqual('');
    })

    it('should work with a name passed', async () => {
      element.match = {
        params: {
          name: 'stencil'
        }
      }
      
      await flush(element);
      expect(element.textContent).toEqual('Hello! My name is stencil. My name was passed in through a route param!');
    });
  });
});